+++ javascript/.prettierrc.json
