+++ javascript/package.json
@@ -1,21 +1,29 @@
 {
-  "name": "demo",
+  "name": "eshopit",
   "version": "1.0.0",
-  "description": "",
+  "description": "Example of an application integrated with Transcend.",
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
+    "dev": "nodemon index.js",
+    "lint": "eslint",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "",
+  "author": "Transcend Inc.",
   "license": "ISC",
   "dependencies": {
-    "body-parser": "^1.18.3",
-    "express": "^4.16.4",
-    "jws": "^4.0.0",
-    "request": "^2.88.0"
+    "body-parser": "^1.19.0",
+    "dotenv": "^8.2.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "express": "^4.17.1",
+    "got": "^11.6.2",
+    "jsonwebtoken": "^8.5.1"
   },
   "devDependencies": {
-    "ngrok": "^3.1.0"
+    "eslint": "^7.9.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "nodemon": "^2.0.4",
+    "prettier": "^2.1.2"
   }
 }
