+++ javascript/util.js
